+++ app/src/main/kotlin/io/github/erikcaffrey/livebinding/di/ServiceLocator.kt
@@ -1,0 +1,11 @@
+package io.github.erikcaffrey.livebinding.di
+
+import io.github.erikcaffrey.livebinding.model.DcCharacterDataSource
+import io.github.erikcaffrey.livebinding.model.DcCharacterRepository
+
+object ServiceLocator {
+
+    private val dcCharacterDataSource: DcCharacterDataSource = DcCharacterDataSource()
+
+    val dcCharacterRepository = DcCharacterRepository(dcCharacterDataSource)
+}
