+++ app/src/main/kotlin/io/github/erikcaffrey/livebinding/model/DcCharacter.kt
@@ -1,3 +1,3 @@
-package io.github.erikcaffrey.livedata_databinding.model
+package io.github.erikcaffrey.livebinding.model
 
 data class DcCharacter(val name: String, val imageProfile: String)
