+++ app/src/main/kotlin/io/github/erikcaffrey/livebinding/view/ImageBindingAdapter.kt
@@ -1,0 +1,11 @@
+package io.github.erikcaffrey.livebinding.view
+
+import android.widget.ImageView
+import androidx.databinding.BindingAdapter
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
+
+@BindingAdapter("imageUrl")
+fun loadImage(imageView: ImageView, url: String) {
+    Glide.with(imageView.context).load(url).transition(DrawableTransitionOptions.withCrossFade()).into(imageView)
+}
