--- app/src/main/kotlin/io/github/erikcaffrey/livedata_databinding/viewmodel/DcCharacterViewModel.kt
@@ -1,20 +1,0 @@
-package io.github.erikcaffrey.livedata_databinding.viewmodel
-
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.ViewModel
-import io.github.erikcaffrey.livedata_databinding.di.ServiceLocator
-import io.github.erikcaffrey.livedata_databinding.model.DcCharacter
-
-class DcCharacterViewModel : ViewModel() {
-
-    private val dcCharacterRepository = ServiceLocator.dcCharacterRepository
-    private val dcCharacterListMutableLive = MutableLiveData<List<DcCharacter>>()
-
-    init {
-        dcCharacterRepository.receiverDcCharacterList { it ->
-            dcCharacterListMutableLive.postValue(it)
-        }
-    }
-
-    fun getDcCharacterList() = dcCharacterListMutableLive
-}
